requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==1.4.47
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
pillow==10.0.1
pandas==2.1.3
requests
python-multipart